refactor: Move API analysis to a central location
(This will allow us to remove obsolete commands that use the API index without causing problems.)

Only the static methods are used in this commit, but they will become instance methods in subsequent commits.
